Fix MSVC warnings.
aspects\qabstractaspect.cpp(50) : warning C4267: 'argument' : conversion from 'size_t' to 'int', possible loss of data
resources\qframeallocator.cpp(276) : warning C4267: 'argument' : conversion from 'size_t' to 'uint', possible loss of data
resources\qframeallocator.cpp(284) : warning C4267: 'argument' : conversion from 'size_t' to 'uint', possible loss of data
qchangearbiter.cpp(108) : warning C4267: 'initializing' : conversion from 'size_t' to 'int', possible loss of data
qt3d/src/render/backend/jobs/renderviewjobutils_p.h(64) : warning C4099: 'Qt3D::Render::ShaderUniform' : type name first seen using 'struct' now seen using 'class'

+++ src/core/resources/qframeallocator.cpp
@@ -273,7 +273,7 @@ void* QFrameAllocator::allocateRawMemory(size_t size)
 {
     Q_D(QFrameAllocator);
     Q_ASSERT(size <= d->m_maxObjectSize);
-    uint allocatorIndex = d->allocatorIndexFromSize(size);
+    uint allocatorIndex = d->allocatorIndexFromSize(uint(size));
     return d->allocateAtChunk(allocatorIndex);
 }
 
@@ -281,7 +281,7 @@ void QFrameAllocator::deallocateRawMemory(void* ptr, size_t size)
 {
     Q_D(QFrameAllocator);
     Q_ASSERT(size <= d->m_maxObjectSize);
-    uint allocatorIndex = d->allocatorIndexFromSize(size);
+    uint allocatorIndex = d->allocatorIndexFromSize(uint(size));
     d->deallocateAtChunck(ptr, allocatorIndex);
 }
